Conform inspectaccrualsnapshot and parseaccrualsnapshotfile
