+++ src/rpcserver.cpp
@@ -348,7 +348,6 @@ static const CRPCCommand vRPCCommands[] =
     { "magnitude",               &magnitude,               cat_mining        },
     { "myneuralhash",            &myneuralhash,            cat_mining        },
     { "neuralhash",              &neuralhash,              cat_mining        },
-    { "parseaccrualsnapshotfile",&parseaccrualsnapshotfile,cat_mining        },
     { "pendingbeaconreport",     &pendingbeaconreport,     cat_mining        },
     { "resetcpids",              &resetcpids,              cat_mining        },
     { "revokebeacon",            &revokebeacon,            cat_mining        },
@@ -374,6 +373,7 @@ static const CRPCCommand vRPCCommands[] =
     { "listprojects",            &listprojects,            cat_developer     },
     { "logging",                 &logging,                 cat_developer     },
     { "network",                 &network,                 cat_developer     },
+    { "parseaccrualsnapshotfile",&parseaccrualsnapshotfile,cat_developer     },
     { "parselegacysb",           &parselegacysb,           cat_developer     },
     { "projects",                &projects,                cat_developer     },
     { "readconfig",              &readconfig,              cat_developer     },
